scenes as classes
