+++ src/main/java/jtp/cc1/ascsp1/MonsterGameServlet.java
@@ -43,90 +43,6 @@ public class MonsterGameServlet extends HttpServlet {
     return "{ \"screen\": \"" + screen + "\", \"method\": \"" + method + "\", \"other\": \"" + other + "\" }";
   }
   
-  private class OopsScene implements SceneObject {
-    private SceneObject back;
-    public String method() {
-      return "read";
-    }
-    public String draw() {
-      return "<br>Que!? ";
-    }
-    public SceneObject whereToNext(String input) {
-      return back;
-    }
-    public OopsScene(SceneObject b) {
-      back = b;
-    }
-  }
-  
-  private class MenuScene implements SceneObject {
-    public String method() {
-      return "read";
-    }
-    public String draw() {
-      return "<br>1. New game<br>2. Save game<br>etc...<br>Enter choice: ";
-    }
-    public SceneObject whereToNext(String input) {
-      switch (input) {
-        case "1":
-          g = new Game();
-          return new GameScene();
-        case "2":
-          return new FilerScene();
-        default:
-          return new OopsScene((SceneObject)this);
-      }
-    }
-  }
-  
-  private class GameScene implements SceneObject {
-    public String method() {
-      return g.method;
-    }
-    public String draw() {
-      if (g.isWon()) {
-        return "<br>WINNER!!!!<br>Press any key to continue: ";
-      } else if (g.isLost()) {
-        return "<br>LOOSER!!!!<br>Press any key to continue: ";
-      } else {
-        return "<br>|------" + g.data + "------|<br>Enter NSEWP: ";
-      }
-    }
-    public SceneObject whereToNext(String input) {
-      switch (g.newState(input)) {
-        case "ispaused": case "isover":
-          return new MenuScene();
-        case "isinplay":
-          return new GameScene();
-        default:
-          return new OopsScene((SceneObject)this);
-      }
-    }
-  }
-
-  private class FilerScene implements SceneObject {
-    public String method() {
-      return "readln";
-    }
-    public String draw() {
-      return "<br>--my saved files--<br>Enter filename: ";
-    }
-    public SceneObject whereToNext(String input) {
-      String newFilerState;
-      if (input.contains(" ")) {
-        newFilerState = "fail";
-      } else {
-        newFilerState = "success";
-      }
-      switch (newFilerState) {
-        case "success":
-          return new MenuScene();
-        default:
-          return new OopsScene((SceneObject)this);
-      }
-    }
-  }
-  
   // dynamic object stuff
   private Game g;
